// frontend/src/types/User.ts
export interface User {
    userId: number;
    name: string;
    phone: string;
    email: string;
    age: number;
    gender: string;
    netflix: number;
    amazonPrime: number;
    disneyPlus: number;
    paramountPlus: number;
    max: number;
    hulu: number;
    appleTvPlus: number;
    peacock: number;
    city: string;
    state: string;
    zip: string;
}